before applying styles and template inheritance
