+++ app/routes.py
@@ -2,7 +2,7 @@ from app import app
 from flask import render_template, flash, request, url_for
 from app.forms import BrowsingForm
 from config import Config
-from try2_list_lookup import get_skills_2
+from main import get_skills
 import os
 
 @app.route('/')
@@ -11,6 +11,7 @@ def index():
 
 @app.route('/upload',methods=['GET','POST'])
 def upload():
+    skills = []
     target = os.path.join(Config.APP_ROOT, 'temp')
     print("target folder", target)
 
@@ -23,6 +24,6 @@ def upload():
         destination = "\\".join([target,filename])
         print("destination folder ", destination)
         file.save(destination)
-        skills = get_skills_2(destination)
+        skills.extend(get_skills(destination))
         print(skills)
     return render_template('completed.html', skills=skills)
