+++ app/templates/upload.html
@@ -5,10 +5,11 @@
     </head>
     
     <body>
-            <form action="{{ url_for('upload') }}" method="POST" enctype="multipart/form-data">
-                <input type="file" name="file" multiple>
-                <input type="submit" value="Upload">
-            </form>    
+        <h2>Select the resume</h2>
+        <form action="{{ url_for('upload') }}" method="POST" enctype="multipart/form-data">
+            <input type="file" name="file" multiple>
+            <input type="submit" value="Get Skills">
+        </form>    
     </body>
 </html>
